refactor(modules): remove use of interfaces
Use Pick<T, keyof T> in lieu of interfaces so that we can properly
define the 1-1 relationship between eg a controller and its service,
while still allowing Jest to provide a mock for testing purposes.
