--- src/server/modules/threat/interfaces/FileTypeFilterService.ts
@@ -1,7 +1,0 @@
-import fileUpload from 'express-fileupload'
-
-export interface FileTypeFilterService {
-  hasAllowedType(file: fileUpload.UploadedFile): Promise<boolean>
-}
-
-export default FileTypeFilterService
