+++ src/server/modules/threat/interfaces/index.ts
@@ -1,5 +1,2 @@
-export { SafeBrowsingRepository } from './SafeBrowsingRepository'
-
-export { FileTypeFilterService } from './FileTypeFilterService'
 export { UrlThreatScanService } from './UrlThreatScanService'
 export { VirusScanService } from './VirusScanService'
