--- src/server/modules/user/interfaces/index.ts
@@ -1,3 +1,0 @@
-export { UrlManagementService } from './UrlManagementService'
-
-export default {}
